+++ src/main/java/elice04_pikacharger/pikacharger/domain/charger/dto/payload/ChargerCreateDto.java
@@ -39,7 +39,7 @@ public class ChargerCreateDto {
     private String content;
 
     @Min(value = 0, message = "0원 이상 입력해주세요")
-    private int personalPrice;
+    private double personalPrice;
 
     private List<ChargerTypeDto> chargerTypeDtoList;
 //    private List<ChargerImage> chargerImageList;
